pyproject.toml
[project]
name = "escape-llm-crawl"
version = "0.1.0"
description = "Make your website escape LLM crawling"
readme = "readme.md"
requires-python = "==3.10.18"
dependencies = [
    "flask==3.1.3",
    "pillow==12.1.1",
    "cryptography==46.0.5",
]

[tool.setuptools]
packages = []

[project.optional-dependencies]
dev = [
    "pytest",
]
